d3d9: fix capture triggering — add Tick() to Present for F12 key detection
Present() was missing the RenderDoc::Inst().Tick() call that polls for
capture key (F12) presses each frame. Without it, TriggerCapture() was
never called, so ShouldTriggerCapture() always returned false and no
capture could be triggered.

Also adds Keyboard::AddInputWindow registration so key detection is
scoped to the game window, AddActiveDriver to register D3D9 as the
active API, and fixes the same issues in the swap chain Present path.

Includes various other D3D9 driver fixes: resource unwrapping, buffer
stubs, shader/query/stateblock improvements, and initstate handling.

+++ renderdoc/driver/d3d9/d3d9_replay.cpp
@@ -814,7 +814,7 @@ void D3D9Replay::GetBufferData(ResourceId buff, uint64_t offset, uint64_t len, b
     HRESULT hr = vb->Lock((UINT)offset, (UINT)len, &data, D3DLOCK_READONLY);
     if(SUCCEEDED(hr) && data)
     {
-      retData.assign((byte *)data, len);
+      retData.assign((byte *)data, (size_t)len);
       vb->Unlock();
     }
     vb->Release();
@@ -834,7 +834,7 @@ void D3D9Replay::GetBufferData(ResourceId buff, uint64_t offset, uint64_t len, b
     HRESULT hr = ib->Lock((UINT)offset, (UINT)len, &data, D3DLOCK_READONLY);
     if(SUCCEEDED(hr) && data)
     {
-      retData.assign((byte *)data, len);
+      retData.assign((byte *)data, (size_t)len);
       ib->Unlock();
     }
     ib->Release();
